# proyecto/animalitos-reactive/animalitos-app/target/openapi.json
{"openapi":"3.1.0","info":{"title":"Animalitos API","description":"API CRUD reactiva para la gestion de animalitos. Contrato publico documentado con OpenAPI 3.","contact":{"name":"Equipo Animalitos","email":"animalitos@curso.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Local"}],"tags":[{"name":"Animalitos","description":"API CRUD de animalitos. Identificadores publicos en formato UUID."}],"paths":{"/api/v1/animalitos/{id}":{"get":{"tags":["Animalitos"],"summary":"Recuperar un animal por id","description":"Devuelve el animal cuyo identificador publico (UUID) coincide con el proporcionado.","operationId":"getAnimal","parameters":[{"name":"id","in":"path","description":"Identificador publico (UUID) del animal","required":true,"schema":{"type":"string"},"example":"5b3f1d2a-7c0d-4d8a-9b8e-2c6c1f0aafee"}],"responses":{"200":{"description":"Animal encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Animal"}}}},"404":{"description":"No existe animal con ese id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"put":{"tags":["Animalitos"],"summary":"Modificar un animal","description":"Modifica especie y edad. El nombre NO se puede modificar.","operationId":"updateAnimal","parameters":[{"name":"id","in":"path","description":"Identificador publico (UUID) del animal a modificar","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModificarAnimalRequest"}}},"required":true},"responses":{"200":{"description":"Animal modificado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Animal"}}}},"400":{"description":"Datos invalidos (validacion)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"No existe animal con ese id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"delete":{"tags":["Animalitos"],"summary":"Eliminar un animal","description":"Elimina el animal por su id publico y devuelve los datos del animal eliminado.","operationId":"deleteAnimal","parameters":[{"name":"id","in":"path","description":"Identificador publico (UUID) del animal a eliminar","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Animal eliminado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Animal"}}}},"404":{"description":"No existe animal con ese id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/animalitos":{"get":{"tags":["Animalitos"],"summary":"Listar todos los animales","description":"Devuelve la coleccion completa. Si no hay animales devuelve lista vacia.","operationId":"getAllAnimales","responses":{"200":{"description":"Coleccion de animales (puede ser vacia)","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Animal"}}}}}}},"post":{"tags":["Animalitos"],"summary":"Crear un animal","description":"Crea un animal nuevo. El nombre debe ser unico (case-insensitive).","operationId":"createAnimal","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CrearAnimalRequest"}}},"required":true},"responses":{"201":{"description":"Animal creado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Animal"}}}},"400":{"description":"Datos invalidos (validacion)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Ya existe un animal con ese nombre","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"ModificarAnimalRequest":{"type":"object","description":"Datos para modificar un animal. El nombre NO se puede cambiar.","properties":{"especie":{"type":"string","description":"Nueva especie del animal.","enum":["PERRO","GATO","HAMSTER","CONEJO","LORO","PEZ","TORTUGA"],"example":"PERRO"},"edad":{"type":"integer","format":"int32","description":"Nueva edad en años (0-100).","example":7,"maximum":100,"minimum":0}},"required":["edad","especie"]},"Animal":{"type":"object","description":"Animal tal y como lo expone la API.","properties":{"id":{"type":"string","description":"Identificador publico (UUID).","example":"5b3f1d2a-7c0d-4d8a-9b8e-2c6c1f0aafee"},"nombre":{"type":"string","description":"Nombre del animal.","example":"Lucas"},"especie":{"type":"string","description":"Especie del animal.","enum":["PERRO","GATO","HAMSTER","CONEJO","LORO","PEZ","TORTUGA"],"example":"GATO"},"edad":{"type":"integer","format":"int32","description":"Edad en años.","example":3,"maximum":100,"minimum":0}}},"ErrorResponse":{"type":"object","description":"Estructura uniforme de error devuelta por la API.","properties":{"timestamp":{"type":"string","format":"date-time","description":"Instante en que se produjo el error (ISO-8601).","example":"2026-04-27T16:00:00Z"},"status":{"type":"integer","format":"int32","description":"Codigo HTTP.","example":404},"error":{"type":"string","description":"Identificador del tipo de error.","example":"ANIMAL_NO_ENCONTRADO"},"mensaje":{"type":"string","description":"Mensaje legible para el cliente.","example":"No existe animal con id ..."},"detalles":{"type":"array","description":"Detalles de validacion campo a campo (puede ser vacio).","items":{"$ref":"#/components/schemas/ValidationErrorDTO"}}}},"ValidationErrorDTO":{"type":"object","properties":{"campo":{"type":"string"},"mensaje":{"type":"string"}}},"CrearAnimalRequest":{"type":"object","description":"Datos para crear un animal nuevo.","properties":{"nombre":{"type":"string","description":"Nombre unico del animal (2-60 caracteres).","example":"Lucas"},"especie":{"type":"string","description":"Especie del animal.","enum":["PERRO","GATO","HAMSTER","CONEJO","LORO","PEZ","TORTUGA"],"example":"GATO"},"edad":{"type":"integer","format":"int32","description":"Edad en años (0-100).","example":3,"maximum":100,"minimum":0}},"required":["edad","especie","nombre"]}}}}